% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataForSingleApp.R
\name{getDataForSingleApp}
\alias{getDataForSingleApp}
\title{getDataForSingleApp}
\usage{
getDataForSingleApp(app_url)
}
\arguments{
\item{app_url}{The url for an app page in the Google Play Store}
}
\value{
data
}
\description{
Scrape an app page in the Google Play Store for information associated with a number of html tags.
}
\examples{
app_url <- "https://play.google.com/store/apps/details?id=com.dinaga.photosecret"
getDataForSingleApp(app_url)
}
