% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataForTag.R
\name{getDataForTag}
\alias{getDataForTag}
\title{getDataForTag}
\usage{
getDataForTag(webpage, html_tag, html_attribute, html_attribute_value)
}
\arguments{
\item{webpage}{The html for an app page in the Google Play Store}

\item{html_tag}{An html tag. E.g. "div"}

\item{html_attribute}{An html attribute associated with the html tag. E.g. "class"}

\item{html_attribute_value}{The value associated with the html attribute. E.g. "KmO8jd"}
}
\value{

}
\description{
Scrape an app page in the Google Play Store for information associated with a specific html tag.
}
\examples{
webpage <- read_html("https://play.google.com/store/apps/details?id=com.dinaga.photosecret")
getDataForTag(webpage,"div","class","KmO8jd")
getDataForTag(webpage,"a","class","hrTbp R8zArc")
}
