% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataForApps.R
\name{getDataForApps}
\alias{getDataForApps}
\title{Scrape Google Play Store search results}
\usage{
getDataForApps(search_term)
}
\arguments{
\item{search_term}{getDataForApps will search the Google Play Store for apps using this search term}
}
\value{
data
}
\description{
Scrape Google Play Store search results
}
\examples{
data <- getDataForApps("steganography")
}
